+++ Code/Editor/Plugins/ComponentEntityEditorPlugin/UI/Outliner/OutlinerListModel.cpp
@@ -2476,10 +2476,10 @@ void OutlinerItemDelegate::paint(QPainter* painter, const QStyleOptionViewItem& 
 
     auto backgroundBoxRect = option.rect;
 
-    backgroundBoxRect.setX(backgroundBoxRect.x() + 0.5);
-    backgroundBoxRect.setY(backgroundBoxRect.y() + 2.5);
-    backgroundBoxRect.setWidth(backgroundBoxRect.width() - 1.0);
-    backgroundBoxRect.setHeight(backgroundBoxRect.height() - 1.0);
+    backgroundBoxRect.setX(static_cast<int>(backgroundBoxRect.x() + 0.5f));
+    backgroundBoxRect.setY(static_cast<int>(backgroundBoxRect.y() + 2.5f));
+    backgroundBoxRect.setWidth(static_cast<int>(backgroundBoxRect.width() - 1.0f));
+    backgroundBoxRect.setHeight(static_cast<int>(backgroundBoxRect.height() - 1.0f));
 
     const qreal sliceBorderHeight = 0.8f;
 
@@ -2513,7 +2513,7 @@ void OutlinerItemDelegate::paint(QPainter* painter, const QStyleOptionViewItem& 
             else
             {
                 auto newRect = option.rect;
-                newRect.setHeight(newRect.height() - 1.0);
+                newRect.setHeight(static_cast<int>(newRect.height() - 1.0f));
                 path.addRect(newRect);
             }
